Expose our universal `Class.from_yaml_file` interface in more places
Specifically, this exposes it from:
- `ElasticGraph::SchemaArtifacts`
- `ElasticGraph::DatastoreCore`

I'll be using this new API in the READMEs of those gems.
