+++ elasticgraph-schema_artifacts/lib/elastic_graph/schema_artifacts/from_disk.rb
@@ -10,35 +10,11 @@ require "elastic_graph/constants"
 require "elastic_graph/errors"
 require "elastic_graph/schema_artifacts/artifacts_helper_methods"
 require "elastic_graph/schema_artifacts/runtime_metadata/schema"
-require "elastic_graph/support/hash_util"
 require "elastic_graph/support/memoizable_data"
 require "yaml"
 
 module ElasticGraph
-  # Namespace for all code related to ElasticGraph schema artifacts.
-  # Schema artifacts are generated by {SchemaDefinition} and dumped to disk.
-  # Later, they are used to power ElasticGraph.
   module SchemaArtifacts
-    # Builds a `SchemaArtifacts::FromDisk` instance using the provided YAML settings.
-    #
-    # @param parsed_yaml [Hash<String, Object>] hash parsed from a settings YAML file
-    # @return [FromDisk]
-    def self.from_parsed_yaml(parsed_yaml)
-      schema_artifacts = parsed_yaml.fetch("schema_artifacts") do
-        raise Errors::ConfigError, "Config is missing required key `schema_artifacts`."
-      end
-
-      if (extra_keys = schema_artifacts.keys - ["directory"]).any?
-        raise Errors::ConfigError, "Config has extra `schema_artifacts` keys: #{extra_keys}"
-      end
-
-      directory = schema_artifacts.fetch("directory") do
-        raise Errors::ConfigError, "Config is missing required key `schema_artifacts.directory`."
-      end
-
-      FromDisk.new(directory)
-    end
-
     # Responsible for loading schema artifacts from disk and providing access to each artifact.
     #
     # @!attribute [r] artifacts_dir
