+++ elasticgraph-schema_artifacts/spec/unit/elastic_graph/schema_artifacts/from_disk_spec.rb
@@ -12,35 +12,6 @@ require "elastic_graph/schema_artifacts/from_disk"
 
 module ElasticGraph
   module SchemaArtifacts
-    RSpec.describe SchemaArtifacts, ".from_parsed_yaml" do
-      it "uses the `schema_artifacts.directory` setting to build a `FromDisk` instance" do
-        artifacts = from_parsed_yaml({"schema_artifacts" => {"directory" => "some_dir"}})
-
-        expect(artifacts).to be_a(FromDisk)
-        expect(artifacts.artifacts_dir).to eq "some_dir"
-      end
-
-      it "fails with a clear error if the required keys are missing" do
-        expect {
-          from_parsed_yaml({})
-        }.to raise_error Errors::ConfigError, a_string_including("schema_artifacts")
-
-        expect {
-          from_parsed_yaml({"schema_artifacts" => {}})
-        }.to raise_error Errors::ConfigError, a_string_including("schema_artifacts.directory")
-      end
-
-      it "fails with a clear error if extra `schema_artifacts` settings are provided" do
-        expect {
-          from_parsed_yaml({"schema_artifacts" => {"directory" => "a", "foo" => 3}})
-        }.to raise_error Errors::ConfigError, a_string_including("foo")
-      end
-
-      def from_parsed_yaml(parsed_yaml)
-        SchemaArtifacts.from_parsed_yaml(parsed_yaml)
-      end
-    end
-
     RSpec.describe FromDisk do
       it "loads each schema artifact from disk" do
         artifacts = FromDisk.new(::File.join(CommonSpecHelpers::REPO_ROOT, "config", "schema", "artifacts"))
